vocsn-report-generator/python_requirements.txt
reportlab>=3.5.23
plotly>=4.1.1
psutil>=5.6.3
pytz>=2019.2
requests>=2.22.0
pandas>=0.25.1
crc16>=0.1.1
azure-cosmosdb-table>=1.0.5
azure-storage-file>=2.1.0